# Insights into performance gains of tuning min/max validators