@stevehu 2019-07-20
@kosty The numbers are very interesting. I always thought integer should be the fastest but it is not the case. I need to dig into it deeper to understand how the library handles these use cases internally. This performance test gives us a lot of insight into how the current implementation behaves and it can be used as a based line to compare with any further enhancements. Thanks a lot for the effort.